requirements.txt
astroid==1.5.3
colorama==0.3.9
Cython==0.25.2
Django==1.11
django-cors-headers==2.0.2
django-filter==1.0.4
djangorestframework==3.6.2
dnspython==1.15.0
isort==4.2.15
lazy-object-proxy==1.3.1
mccabe==0.6.1
mysqlclient==1.3.10
olefile==0.44
Pillow==4.1.1
pylint==1.7.1
pytz==2017.2
requests==2.14.2
virtualenv==15.1.0
virtualenvwrapper-win==1.2.1
whitenoise==3.3.0
wrapt==1.10.10
psycopg2>=2.7,<2.8 --no-binary :all:
